requirements.txt
Flask
Flask-Bcrypt
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Mail
python-dotenv
gunicorn
psycopg2-binary